pyproject.toml
[project]
name = "notebook-lm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "pymupdf>=1.26.7",
    "fastembed==0.7.3",
    "litai>=0.0.10",
    "mistralai>=1.10.1",
    "python-dotenv>=1.0.0",
    "zep-cloud==3.4.3",
    "crewai>=1.6.1",
    "zep-crewai==1.1.1",
    "firecrawl-py==4.3.6",
    "yt-dlp>=2024.12.13",
    "assemblyai==0.44.3",
]
